+++ NitroPaint/ncerviewer.h
@@ -90,6 +90,6 @@ HWND CreateNcerViewerImmediate(int x, int y, int width, int height, HWND hWndPar
 
 void CellViewerRenderGridlines(FrameBuffer *fb, int scale, int scrollX, int scrollY);
 
-void CellViewerRenderCell(COLOR32 *px, int *covbuf, NCER *ncer, NCGR *ncgr, NCLR *nclr, int cellIndex, NCER_CELL *cell, int xOffs, int yOffs, float a, float b, float c, float d);
+void CellViewerRenderCell(COLOR32 *px, int *covbuf, NCER *ncer, NCGR *ncgr, NCLR *nclr, int cellIndex, NCER_CELL *cell, int xOffs, int yOffs, double a, double b, double c, double d);
 
 COLOR32 *CellViewerCropRenderedCell(COLOR32 *px, int width, int height, int *pMinX, int *pMinY, int *outWidth, int *outHeight);
